requirements.txt
fastapi
uvicorn
databases
pydantic
python-jose
httpx
supabase
python-dotenv
python-multipart
aiosmtplib
fpdf2
